server: Extract route functions and Filter generation code into files
